Move common datetime functions to DateTimeExtractorUtils
Will be used for en-us
