+++ numbers/src/main/java/org/dicio/numbers/lang/it/ItalianDateTimeExtractor.java
@@ -1,5 +1,7 @@
 package org.dicio.numbers.lang.it;
 
+import static org.dicio.numbers.util.DateTimeExtractorUtils.HOURS_IN_DAY;
+import static org.dicio.numbers.util.DateTimeExtractorUtils.isMomentOfDayPm;
 import static org.dicio.numbers.util.NumberExtractorUtils.signBeforeNumber;
 import static org.dicio.numbers.util.Utils.firstNotNull;
 import static org.dicio.numbers.util.Utils.roundToInt;
@@ -7,34 +9,29 @@ import static org.dicio.numbers.util.Utils.roundToInt;
 import org.dicio.numbers.parser.lexer.TokenStream;
 import org.dicio.numbers.unit.Duration;
 import org.dicio.numbers.unit.Number;
+import org.dicio.numbers.util.DateTimeExtractorUtils;
 import org.dicio.numbers.util.DurationExtractorUtils;
 import org.dicio.numbers.util.NumberExtractorUtils;
 
-import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
-import java.time.Month;
 import java.time.temporal.ChronoUnit;
-import java.util.function.Function;
-import java.util.function.Supplier;
 
 public class ItalianDateTimeExtractor {
-
-    private static final int HOURS_IN_DAY = 24;
-    private static final int DAYS_IN_WEEK = DayOfWeek.values().length; // 7
-    private static final int MONTHS_IN_YEAR = Month.values().length; // 12
 
     private final TokenStream ts;
     private final LocalDateTime now;
     private final ItalianNumberExtractor numberExtractor;
     private final DurationExtractorUtils durationExtractor;
+    public final DateTimeExtractorUtils dateTimeExtractor;
 
     ItalianDateTimeExtractor(final TokenStream tokenStream, final LocalDateTime now) {
         this.ts = tokenStream;
         this.now = now;
         this.numberExtractor = new ItalianNumberExtractor(ts);
         this.durationExtractor = new DurationExtractorUtils(ts, numberExtractor::numberNoOrdinal);
+        this.dateTimeExtractor = new DateTimeExtractorUtils(ts, now, this::extractIntegerInRange);
     }
 
     private Integer extractIntegerInRange(final int fromInclusive, final int toInclusive) {
@@ -123,8 +120,8 @@ public class ItalianDateTimeExtractor {
             }
         } else {
             // found a time, now look for am/pm or a moment of day
-            pm = ts.tryOrSkipDateTimeIgnore(true,
-                    () -> firstNotNull(this::ampm, () -> isMomentOfDayPm(momentOfDay())));
+            pm = ts.tryOrSkipDateTimeIgnore(true, () -> firstNotNull(
+                    dateTimeExtractor::ampm, () -> isMomentOfDayPm(momentOfDay())));
         }
 
         if (pm != null && pm && !isMomentOfDayPm(time.getHour())) {
@@ -143,13 +140,13 @@ public class ItalianDateTimeExtractor {
         LocalTime result = LocalTime.of(hour, 0);
 
         final Integer minute = ts.tryOrSkipDateTimeIgnore(true,
-                () -> firstNotNull(this::specialMinute, this::minute));
+                () -> firstNotNull(this::specialMinute, dateTimeExtractor::minute));
         if (minute == null) {
             return result;
         }
         result = result.withMinute(minute);
 
-        final Integer second = ts.tryOrSkipDateTimeIgnore(true, this::second);
+        final Integer second = ts.tryOrSkipDateTimeIgnore(true, dateTimeExtractor::second);
         if (second == null) {
             return result;
         }
@@ -194,120 +191,32 @@ public class ItalianDateTimeExtractor {
             return null;
         }
 
-        final Boolean bcad = bcad();
+        final Boolean bcad = dateTimeExtractor.bcad();
         return result.withYear(year * (bcad == null || bcad ? 1 : -1));
     }
 
 
-    Integer monthName() {
-        if (ts.get(0).hasCategory("month_name")) {
-            ts.movePositionForwardBy(1);
-            return (int) ts.get(-1).getNumber().integerValue();
-        } else {
-            return null;
-        }
-    }
-
     Integer dayOfWeek() {
-        if (ts.get(0).hasCategory("day_of_week")) {
-            ts.movePositionForwardBy(1);
-            return (int) ts.get(-1).getNumber().integerValue();
-        } else if (ts.get(0).isValue("mar")) {
+        if (ts.get(0).isValue("mar")) {
             ts.movePositionForwardBy(1);
             return 1; // special case, since mar already used for march
         } else {
-            return null;
+            return dateTimeExtractor.dayOfWeek();
         }
     }
 
-
-    Boolean ampm() {
-        return bcadOrAmpm("ampm");
-    }
-
-    Boolean bcad() {
-        return bcadOrAmpm("bcad");
-    }
-
-    /**
-     * @param prefix either "bcad" or "ampm", i.e. the prefix to use for the following categories:
-     *               _before, _after, _identifier, _before_combined, _after_combined
-     * @return false if before+identifier matches, true if after+identifier matches, null otherwise
-     */
-    private Boolean bcadOrAmpm(final String prefix) {
-        final boolean result;
-        ts.movePositionForwardBy(1);
-        if (ts.get(-1).hasCategory(prefix + "_before")) {
-            result = false;
-        } else if (ts.get(-1).hasCategory(prefix + "_after")) {
-            result = true;
-        } else if (ts.get(-1).hasCategory(prefix + "_before_combined")) {
-            // found am or bc in a single word -> return "before"
-            return false;
-        } else if (ts.get(-1).hasCategory(prefix + "_after_combined")) {
-            // found pm or ad in a single word -> return "after"
-            return true;
-        } else {
-            // nothing related to bc/ad/am/pm here (even if the identifier might match on its own)
-            ts.movePositionForwardBy(-1);
-            return null;
-        }
-
-        // we can't use ts.indexOfWithoutCategory, since some ignore words might be identifiers
-        final Boolean foundIdentifier = ts.tryOrSkipCategory("ignore", true,
-                () -> ts.get(0).hasCategory(prefix + "_identifier") ? true : null);
-        if (foundIdentifier != null) {
-            ts.movePositionForwardBy(1);
-            return result;
-        } else {
-            ts.movePositionForwardBy(-1);
-            return null;
-        }
-    }
-
-
-    Integer second() {
-        return minuteOrSecond("1 SECONDS");
-    }
-
-    Integer specialMinute() {
+    public Integer specialMinute() {
         final int originalPosition = ts.getPosition();
 
         final Number number = numberExtractor.numberNoOrdinal();
         if (number != null && number.isDecimal()
                 && number.decimalValue() > 0.0 && number.decimalValue() < 1.0) {
+            // e.g. alle due e tre quarti
             return roundToInt(number.decimalValue() * 60);
         }
 
         ts.setPosition(originalPosition);
         return null;
-    }
-
-    Integer minute() {
-        return minuteOrSecond("1 MINUTES");
-    }
-
-    private Integer minuteOrSecond(final String durationCategory) {
-        final Integer number = extractIntegerInRange(0, 59);
-        if (number == null) {
-            return null;
-        }
-
-        if (ts.get(0).isDurationToken()
-                && ts.get(0).asDurationToken().getDurationCategory().equals(durationCategory)) {
-            // skip "minuti"/"secondi" said after a minute/second count, e.g. ventiquattro minuti
-            ts.movePositionForwardBy(1);
-        }
-
-        return number;
-    }
-
-
-    static Boolean isMomentOfDayPm(final Integer momentOfDay) {
-        if (momentOfDay == null) {
-            return null;
-        }
-        return momentOfDay >= 12;
     }
 
     Integer noonMidnightLike() {
@@ -377,8 +286,8 @@ public class ItalianDateTimeExtractor {
 
 
     private LocalDate relativeSpecialDay() {
-        final Integer days = firstNotNull(this::relativeYesterday, this::relativeToday,
-                this::relativeTomorrow, this::relativeDayOfWeekDuration);
+        final Integer days = firstNotNull(this::relativeYesterday, dateTimeExtractor::relativeToday,
+                this::relativeTomorrow, dateTimeExtractor::relativeDayOfWeekDuration);
         if (days == null) {
             return null;
         }
@@ -414,15 +323,6 @@ public class ItalianDateTimeExtractor {
         return -dayCount;
     }
 
-    Integer relativeToday() {
-        if (ts.get(0).hasCategory("today")) {
-            ts.movePositionForwardBy(1);
-            return 0; // no offset
-        } else {
-            return null;
-        }
-    }
-
     Integer relativeTomorrow() {
         int originalPosition = ts.getPosition();
 
@@ -445,102 +345,8 @@ public class ItalianDateTimeExtractor {
         return dayCount;
     }
 
-    Integer relativeDayOfWeekDuration() {
-        return relativeIndicatorDuration(() -> {
-            Integer number = extractIntegerInRange(1, Integer.MAX_VALUE);
-            if (number == null) {
-                // there does not need to be a number, e.g. giovedì prossimo
-                number = 1;
-            } else {
-                // found a number, e.g. fra due
-                ts.movePositionForwardBy(ts.indexOfWithoutCategory("date_time_ignore", 0));
-            }
-
-            if (ts.get(0).hasCategory("day_of_week")) {
-                // found a day of week, e.g. giovedì
-                final int daysDifference
-                        = (int) ts.get(0).getNumber().integerValue() - now.getDayOfWeek().ordinal();
-                final int daysOffset = (daysDifference + DAYS_IN_WEEK) % DAYS_IN_WEEK
-                        // add a week if the two days coincide
-                        + (daysDifference == 0 ? DAYS_IN_WEEK : 0)
-                        // sum some additional weeks if the input says so
-                        + (number - 1) * DAYS_IN_WEEK;
-                ts.movePositionForwardBy(1);
-                return daysOffset;
-            } else {
-                return null;
-            }
-
-        }, daysOffset -> daysOffset % DAYS_IN_WEEK == 0
-                // the congruency modulo DAYS_IN_WEEK is 0: just use a minus to maintain it
-                ? -daysOffset
-                // keep congruency modulo DAYS_IN_WEEK, taking care of additional weeks
-                : 2 * (daysOffset % DAYS_IN_WEEK) - DAYS_IN_WEEK - daysOffset);
-    }
-
-    Duration relativeMonthDuration() {
-        return relativeIndicatorDuration(() -> {
-            if (ts.get(0).hasCategory("month_name")) {
-                final long monthsDifference
-                        = ts.get(0).getNumber().integerValue() - now.getMonth().ordinal();
-                final long monthsOffset = (monthsDifference + MONTHS_IN_YEAR) % MONTHS_IN_YEAR
-                        // add a year if the two months coincide
-                        + (monthsDifference == 0 ? MONTHS_IN_YEAR : 0);
-                ts.movePositionForwardBy(1);
-                return new Duration(0, 0, monthsOffset, 0);
-            }
-            return null;
-
-        }, duration -> {
-            final long monthsOffset = duration.getMonths();
-            final long newMonthsOffset = monthsOffset == MONTHS_IN_YEAR
-                    // the congruency modulo MONTHS_IN_YEAR is 0: just use a minus to maintain it
-                    ? -MONTHS_IN_YEAR
-                    // keep congruency modulo MONTHS_IN_YEAR
-                    : monthsOffset - MONTHS_IN_YEAR;
-            return new Duration(0, 0, newMonthsOffset, 0);
-        });
-    }
-
     Duration relativeDuration() {
-        return relativeIndicatorDuration(durationExtractor::duration,
+        return dateTimeExtractor.relativeIndicatorDuration(durationExtractor::duration,
                 duration -> duration.multiply(-1));
-    }
-
-    private <T> T relativeIndicatorDuration(final Supplier<T> durationExtractor,
-                                            final Function<T, T> oppositeDuration) {
-        final int originalTsPosition = ts.getPosition();
-
-        int relativeIndicator = 0; // 0 = not found, otherwise the sign, +1 or -1
-        if (ts.get(0).hasCategory("pre_relative_indicator")) {
-            // there is a relative indicator before, e.g. fra
-            relativeIndicator = ts.get(0).hasCategory("negative") ? -1 : 1;
-            ts.movePositionForwardBy(ts.indexOfWithoutCategory("date_time_ignore", 1));
-        }
-
-        final T result = durationExtractor.get();
-        if (result == null) {
-            // no duration found, e.g. tra sei ciao
-            ts.setPosition(originalTsPosition);
-            return null;
-        }
-
-        final int nextNotIgnore = ts.indexOfWithoutCategory("date_time_ignore", 0);
-        if (relativeIndicator == 0 && ts.get(nextNotIgnore).hasCategory("post_relative_indicator")) {
-            // there is a relative indicator after, e.g. due settimane fa
-            relativeIndicator = ts.get(nextNotIgnore).hasCategory("negative") ? -1 : 1;
-            ts.movePositionForwardBy(nextNotIgnore + 1);
-        }
-
-        if (relativeIndicator == 0) {
-            // no relative indicator found, this is not a relative duration, e.g. sei mesi
-            ts.setPosition(originalTsPosition);
-            return null;
-        } else {
-            // found relative duration, e.g. tra due minuti
-            return relativeIndicator == -1
-                    ? oppositeDuration.apply(result)
-                    : result;
-        }
     }
 }
